And the changes
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/branch/EtaewMajor@1125 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/behaviour/Actions/TakeItemAction.cs
@@ -110,7 +110,7 @@ namespace DOL.GS.Behaviour.Actions
                     }
                     else
                     {
-                        playerInventory.RemoveCountFromStack((InventoryItem)de.Key, (int)de.Value);
+                        playerInventory.RemoveCountFromStack((InventoryItem)de.Key, (byte)de.Value);
                     }
                 }
                 playerInventory.CommitChanges();
