+++ GameServer/gameobjects/Bonedancer/BDPet.cs
@@ -305,22 +305,22 @@ namespace DOL.GS
 				InventoryItem item;
 				if ((item = Inventory.GetItem(eInventorySlot.TwoHandWeapon)) != null)
 				{
-					item.DPS_AF = (int)(Level * 3.3);
+					item.DPS_AF = (byte)(Level * 3.3);
 					item.SPD_ABS = 50;
 				}
 				if ((item = Inventory.GetItem(eInventorySlot.RightHandWeapon)) != null)
 				{
-					item.DPS_AF = (int)(Level * 3.3);
+					item.DPS_AF = (byte)(Level * 3.3);
 					item.SPD_ABS = 37;
 				}
 				if ((item = Inventory.GetItem(eInventorySlot.LeftHandWeapon)) != null)
 				{
-					item.DPS_AF = (int)(Level * 3.3);
+					item.DPS_AF = (byte)(Level * 3.3);
 					item.SPD_ABS = 50;
 				}
 				if ((item = Inventory.GetItem(eInventorySlot.DistanceWeapon)) != null)
 				{
-					item.DPS_AF = (int)(Level * 3.3);
+					item.DPS_AF = (byte)(Level * 3.3);
 					item.SPD_ABS = 50;
 					SwitchWeapon(eActiveWeaponSlot.Distance);
 					UpdateNPCEquipmentAppearance();
