+++ GameServer/packets/Client/168/PlayerMoveItemRequestHandler.cs
@@ -27,52 +27,52 @@ using System.Collections.Generic;
 
 namespace DOL.GS.PacketHandler.Client.v168
 {
-	[PacketHandlerAttribute(PacketHandlerType.TCP,0x75^168,"Player move item")]
+	[PacketHandlerAttribute(PacketHandlerType.TCP, 0x75 ^ 168, "Player move item")]
 	public class PlayerMoveItemRequestHandler : IPacketHandler
 	{
 		private static readonly ILog log = LogManager.GetLogger(MethodBase.GetCurrentMethod().DeclaringType);
 
 		public int HandlePacket(GameClient client, GSPacketIn packet)
 		{
-			ushort id		= (ushort) packet.ReadShort();
-			ushort toSlot	= (ushort) packet.ReadShort();
-			ushort fromSlot = (ushort) packet.ReadShort();
-			ushort itemCount= (ushort) packet.ReadShort();
+			ushort id = (ushort)packet.ReadShort();
+			ushort toSlot = (ushort)packet.ReadShort();
+			ushort fromSlot = (ushort)packet.ReadShort();
+			byte itemCount = (byte)packet.ReadShort();
 
 			//log.Debug("MoveItem, id=" + id.ToString() + " toSlot=" + toSlot.ToString() + " fromSlot=" + fromSlot.ToString() + " itemCount=" + itemCount.ToString());
 
 			//If our slot is > 1000 it is (objectID+1000) of target
-			if(toSlot>1000)
+			if (toSlot > 1000)
 			{
-				ushort objectID = (ushort)(toSlot-1000);
-				GameObject obj = WorldMgr.GetObjectByIDFromRegion(client.Player.CurrentRegionID,objectID);
-				if(obj==null || obj.ObjectState!=GameObject.eObjectState.Active)
+				ushort objectID = (ushort)(toSlot - 1000);
+				GameObject obj = WorldMgr.GetObjectByIDFromRegion(client.Player.CurrentRegionID, objectID);
+				if (obj == null || obj.ObjectState != GameObject.eObjectState.Active)
 				{
-					client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
-					client.Out.SendMessage("Invalid trade target. ("+objectID+")", eChatType.CT_System, eChatLoc.CL_SystemWindow);
+					client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
+					client.Out.SendMessage("Invalid trade target. (" + objectID + ")", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 					return 0;
 				}
 
 				GamePlayer tradeTarget = null;
 				//If our target is another player we set the tradetarget
-				if(obj is GamePlayer)
+				if (obj is GamePlayer)
 				{
-					tradeTarget = (GamePlayer) obj;
-					if(tradeTarget.Client.ClientState != GameClient.eClientState.Playing)
+					tradeTarget = (GamePlayer)obj;
+					if (tradeTarget.Client.ClientState != GameClient.eClientState.Playing)
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
-						client.Out.SendMessage("Can't trade with inactive players.", eChatType.CT_System,  eChatLoc.CL_SystemWindow);
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
+						client.Out.SendMessage("Can't trade with inactive players.", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 						return 0;
 					}
-					if(tradeTarget == client.Player)
+					if (tradeTarget == client.Player)
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
-						client.Out.SendMessage("You can't trade with yourself, silly!",eChatType.CT_System,eChatLoc.CL_SystemWindow);
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
+						client.Out.SendMessage("You can't trade with yourself, silly!", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 						return 0;
 					}
-					if(!GameServer.ServerRules.IsAllowedToTrade(client.Player, tradeTarget, false))
+					if (!GameServer.ServerRules.IsAllowedToTrade(client.Player, tradeTarget, false))
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 						return 0;
 					}
 				}
@@ -135,56 +135,56 @@ namespace DOL.GS.PacketHandler.Client.v168
 				}
 
 				//Is the "item" we want to move money? For Version 1.78+
-				if(client.Version >= GameClient.eClientVersion.Version178
+				if (client.Version >= GameClient.eClientVersion.Version178
 				   && fromSlot >= (int)eInventorySlot.Mithril178 && fromSlot <= (int)eInventorySlot.Copper178)
 				{
 					fromSlot -= eInventorySlot.Mithril178 - eInventorySlot.Mithril;
 				}
 
 				//Is the "item" we want to move money?
-				if(fromSlot>=(ushort)eInventorySlot.Mithril && fromSlot<=(ushort)eInventorySlot.Copper)
+				if (fromSlot >= (ushort)eInventorySlot.Mithril && fromSlot <= (ushort)eInventorySlot.Copper)
 				{
-					int[] money=new int[5];
-					money[fromSlot-(ushort)eInventorySlot.Mithril]=itemCount;
-					long flatmoney = Money.GetMoney(money[0],money[1],money[2],money[3],money[4]);
+					int[] money = new int[5];
+					money[fromSlot - (ushort)eInventorySlot.Mithril] = itemCount;
+					long flatmoney = Money.GetMoney(money[0], money[1], money[2], money[3], money[4]);
 
-					if(client.Version >= GameClient.eClientVersion.Version178) // add it back for proper slot update...
+					if (client.Version >= GameClient.eClientVersion.Version178) // add it back for proper slot update...
 						fromSlot += eInventorySlot.Mithril178 - eInventorySlot.Mithril;
 
-					if(!WorldMgr.CheckDistance(obj, client.Player, WorldMgr.GIVE_ITEM_DISTANCE))
+					if (!WorldMgr.CheckDistance(obj, client.Player, WorldMgr.GIVE_ITEM_DISTANCE))
 					{
-                        if (obj is GamePlayer)
-                            client.Out.SendMessage(LanguageMgr.GetTranslation(client, "PlayerMoveItemRequestHandler.TooFarAway", client.Player.GetName((GamePlayer)obj)), eChatType.CT_System, eChatLoc.CL_SystemWindow);
-                        else
-                            client.Out.SendMessage(LanguageMgr.GetTranslation(client, "PlayerMoveItemRequestHandler.TooFarAway", obj.GetName(0, false)), eChatType.CT_System, eChatLoc.CL_SystemWindow);
+						if (obj is GamePlayer)
+							client.Out.SendMessage(LanguageMgr.GetTranslation(client, "PlayerMoveItemRequestHandler.TooFarAway", client.Player.GetName((GamePlayer)obj)), eChatType.CT_System, eChatLoc.CL_SystemWindow);
+						else
+							client.Out.SendMessage(LanguageMgr.GetTranslation(client, "PlayerMoveItemRequestHandler.TooFarAway", obj.GetName(0, false)), eChatType.CT_System, eChatLoc.CL_SystemWindow);
 
 						//client.Out.SendMessage("You are too far away to give anything to " + obj.GetName(0, false) + ".", eChatType.CT_System, eChatLoc.CL_SystemWindow);
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 						return 0;
 					}
 
-					if(flatmoney > client.Player.GetCurrentMoney())
+					if (flatmoney > client.Player.GetCurrentMoney())
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 						return 0;
 					}
 
 					client.Player.Notify(GamePlayerEvent.GiveMoney, client.Player, new GiveMoneyEventArgs(client.Player, obj, flatmoney));
 
-					if(tradeTarget!=null)
+					if (tradeTarget != null)
 					{
 						tradeTarget.ReceiveTradeMoney(client.Player, flatmoney);
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 						return 1;
 					}
 
-					if(obj.ReceiveMoney(client.Player, flatmoney))
+					if (obj.ReceiveMoney(client.Player, flatmoney))
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 						return 0;
 					}
 
-					client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+					client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 					return 0;
 				}
 
@@ -208,44 +208,44 @@ namespace DOL.GS.PacketHandler.Client.v168
 					return 0;
 				}
 
-				houseVault.MoveItem(client.Player.Inventory, (eInventorySlot)fromSlot, 
+				houseVault.MoveItem(client.Player.Inventory, (eInventorySlot)fromSlot,
 					(eInventorySlot)toSlot);
 
 				return 1;
 			}
 
 			//Do we want to move an item from inventory/vault/quiver into inventory/vault/quiver?
-			if (((fromSlot>=(ushort)eInventorySlot.Ground && fromSlot<=(ushort)eInventorySlot.LastBackpack)
-				|| (fromSlot>=(ushort)eInventorySlot.FirstVault && fromSlot<=(ushort)eInventorySlot.LastVault))
-				&&((toSlot>=(ushort)eInventorySlot.Ground && toSlot<=(ushort)eInventorySlot.LastBackpack)
-				|| (toSlot>=(ushort)eInventorySlot.FirstVault && toSlot<=(ushort)eInventorySlot.LastVault))
-				|| (toSlot>=(ushort)eInventorySlot.HousingInventory_First && toSlot<=(ushort)eInventorySlot.HousingInventory_Last))
+			if (((fromSlot >= (ushort)eInventorySlot.Ground && fromSlot <= (ushort)eInventorySlot.LastBackpack)
+				|| (fromSlot >= (ushort)eInventorySlot.FirstVault && fromSlot <= (ushort)eInventorySlot.LastVault))
+				&& ((toSlot >= (ushort)eInventorySlot.Ground && toSlot <= (ushort)eInventorySlot.LastBackpack)
+				|| (toSlot >= (ushort)eInventorySlot.FirstVault && toSlot <= (ushort)eInventorySlot.LastVault))
+				|| (toSlot >= (ushort)eInventorySlot.HousingInventory_First && toSlot <= (ushort)eInventorySlot.HousingInventory_Last))
 			{
 				//We want to drop the item
-				if (toSlot==(ushort)eInventorySlot.Ground)
+				if (toSlot == (ushort)eInventorySlot.Ground)
 				{
 					InventoryItem item = client.Player.Inventory.GetItem((eInventorySlot)fromSlot);
 					if (item == null)
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
-						client.Out.SendMessage("Invalid item (slot# "+fromSlot+").",eChatType.CT_System,eChatLoc.CL_SystemWindow);
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
+						client.Out.SendMessage("Invalid item (slot# " + fromSlot + ").", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 						return 0;
 					}
-					if (fromSlot<(ushort)eInventorySlot.FirstBackpack)
+					if (fromSlot < (ushort)eInventorySlot.FirstBackpack)
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 						return 0;
 					}
-					if(!item.IsDropable)
+					if (!item.IsDropable)
 					{
-						client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
-						client.Out.SendMessage("You can not drop this item!",eChatType.CT_System,eChatLoc.CL_SystemWindow);
+						client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
+						client.Out.SendMessage("You can not drop this item!", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 						return 0;
 					}
 
 					if (client.Player.DropItem((eInventorySlot)fromSlot))
 					{
-						client.Out.SendMessage("You drop " + item.GetName(0, false) + " on the ground!",eChatType.CT_System,eChatLoc.CL_SystemWindow);
+						client.Out.SendMessage("You drop " + item.GetName(0, false) + " on the ground!", eChatType.CT_System, eChatLoc.CL_SystemWindow);
 						return 1;
 					}
 					client.Out.SendInventoryItemsUpdate(null);
@@ -258,39 +258,39 @@ namespace DOL.GS.PacketHandler.Client.v168
 			}
 
 
-			if (((fromSlot>=(ushort)eInventorySlot.Ground && fromSlot<=(ushort)eInventorySlot.LastBackpack)
-				|| (fromSlot>=(ushort)eInventorySlot.FirstVault && fromSlot<=(ushort)eInventorySlot.LastVault))
-				&& (toSlot==(ushort)eInventorySlot.PlayerPaperDoll || toSlot==(ushort)eInventorySlot.NewPlayerPaperDoll))
+			if (((fromSlot >= (ushort)eInventorySlot.Ground && fromSlot <= (ushort)eInventorySlot.LastBackpack)
+				|| (fromSlot >= (ushort)eInventorySlot.FirstVault && fromSlot <= (ushort)eInventorySlot.LastVault))
+				&& (toSlot == (ushort)eInventorySlot.PlayerPaperDoll || toSlot == (ushort)eInventorySlot.NewPlayerPaperDoll))
 			{
 				InventoryItem item = client.Player.Inventory.GetItem((eInventorySlot)fromSlot);
-				if(item==null) return 0;
+				if (item == null) return 0;
 
-				toSlot=0;
-				if(item.Item_Type >= (int)eInventorySlot.MinEquipable &&
+				toSlot = 0;
+				if (item.Item_Type >= (int)eInventorySlot.MinEquipable &&
 					item.Item_Type <= (int)eInventorySlot.MaxEquipable)
-					toSlot = (ushort)item.Item_Type;						
-				if (toSlot==0)
+					toSlot = (ushort)item.Item_Type;
+				if (toSlot == 0)
 				{
-					client.Out.SendInventorySlotsUpdate(new int[] {fromSlot});
+					client.Out.SendInventorySlotsUpdate(new int[] { fromSlot });
 					return 0;
 				}
-				if( toSlot == (int)eInventorySlot.LeftBracer || toSlot == (int)eInventorySlot.RightBracer)
+				if (toSlot == (int)eInventorySlot.LeftBracer || toSlot == (int)eInventorySlot.RightBracer)
 				{
-					if(client.Player.Inventory.GetItem(eInventorySlot.LeftBracer) == null)
+					if (client.Player.Inventory.GetItem(eInventorySlot.LeftBracer) == null)
 						toSlot = (int)eInventorySlot.LeftBracer;
 					else
 						toSlot = (int)eInventorySlot.RightBracer;
 				}
 
-				if( toSlot == (int)eInventorySlot.LeftRing || toSlot == (int)eInventorySlot.RightRing)
+				if (toSlot == (int)eInventorySlot.LeftRing || toSlot == (int)eInventorySlot.RightRing)
 				{
-					if(client.Player.Inventory.GetItem(eInventorySlot.LeftRing) == null)
+					if (client.Player.Inventory.GetItem(eInventorySlot.LeftRing) == null)
 						toSlot = (int)eInventorySlot.LeftRing;
 					else
 						toSlot = (int)eInventorySlot.RightRing;
 				}
 
-				client.Player.Inventory.MoveItem((eInventorySlot)fromSlot,(eInventorySlot)toSlot, itemCount);
+				client.Player.Inventory.MoveItem((eInventorySlot)fromSlot, (eInventorySlot)toSlot, itemCount);
 				return 1;
 			}
 			client.Out.SendInventoryItemsUpdate(null);
